Move trafficmgr/service_proxy.go to daemon/mgrproxy.go
The managerproxy has the same lifecycle as the service, not the session,
and this enables a more private initialization of it.
